% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSRTcalc-package.R
\docType{package}
\name{SSRTcalc-package}
\alias{SSRTcalc}
\alias{SSRTcalc-package}
\title{SSRTcalc: Stop-Signal Reaction Time Calculator with Monte Carlo and
Bayesian Extensions}
\description{
Estimates stop-signal reaction time (SSRT) using the integration and mean
methods of Verbruggen et al. (2019), and extends these point estimates
with three families of tools:
}
\details{
\itemize{
  \item \strong{Monte Carlo} (\code{\link{ssrt_boot}}, \code{\link{ssrt_simulate}},
    \code{\link{ssrt_power}}, \code{\link{ssrt_robustness}}): nonparametric
    bootstrap confidence intervals, parametric ex-Gaussian simulation,
    minimum-trial-count / power analysis, and sensitivity of SSRT to
    violations of the horse-race assumptions.
  \item \strong{Bayesian / Stan} (\code{\link{ssrt_stan}} and friends):
    single-subject and hierarchical ex-Gaussian horse-race models fit via
    Hamiltonian Monte Carlo, with an optional trigger-failure parameter
    following Matzke et al. (2013), posterior inhibition functions, and
    posterior predictive checks. Works with either the \pkg{cmdstanr} or
    \pkg{rstan} backend.
  \item \code{\link{run_all_mc}} and \code{\link{ssrt_stan_compare}}:
    convenience wrappers that run a full battery of analyses in one call.
}
}
\section{Typical workflow}{

\preformatted{
data(adaptive)
d <- adaptive[adaptive$SubjID == 1, ]

integration_adaptiveSSD(d)      # point estimate
ssrt_boot(d)                    # + bootstrap CI
ssrt_stan(d)                    # + full posterior (requires cmdstanr/rstan)
}
}

\references{
Verbruggen, F., Aron, A. R., Band, G. P. H., Beste, C., Bissett, P. G.,
Brockett, A. T., ... Boehler, C. N. (2019). A consensus guide to capturing
the ability to inhibit actions and impulses: the stop-signal task.
\emph{eLife}, 8, e46323. \doi{10.7554/eLife.46323}

Matzke, D., Dolan, C. V., Logan, G. D., Brown, S. D., & Wagenmakers, E.-J.
(2013). Bayesian parametric estimation of stop-signal reaction time
distributions. \emph{Journal of Experimental Psychology: General}, 142(4),
1047--1073. \doi{10.1037/a0030543}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agleontyev/SSRTcalc}
  \item Report bugs at \url{https://github.com/agleontyev/SSRTcalc/issues}
}

}
\author{
\strong{Maintainer}: Anton Leontyev \email{anton.leontyev@example.com} (\href{https://orcid.org/0000-0001-8880-6161}{ORCID})

}
\keyword{internal}
